Ender 3 S1 Pro Support/F4 Updates
Ender 3 S1 Pro added for new thermistor type #13 and high temp setting pre-setup.

F4 board compile is building correctly but some boards are not flashing. This may be an issue with the bootloader on the board itself and not a code issue as none of our F4 boards are taking any firmware, including firmware from Creality themselves.